# Refine the timing of PaddingRNN and optimize the infer implementation.