--- a/PaddingRNN/lstm_paddle/args.py
+++ b/PaddingRNN/lstm_paddle/args.py
@@ -45,6 +45,14 @@ def parse_args():
     parser.add_argument(
         '--log_path',
         help='path of the log file. If not set, logs are printed to console')
+    parser.add_argument(
+        '--save_model_dir',
+        type=str, default="models", help='dir of the saved model.')
+    parser.add_argument(
+        '--init_model', type=str, help='path of the init model.')
     parser.add_argument('--enable_ce', action='store_true')
+    parser.add_argument(
+        "--profile", type=bool, default=False, help="whether enable the profiler.")
+    parser.add_argument('--batch_size', type=int, default=0, help='batch size')
     args = parser.parse_args()
     return args