In the current implementation of PaddingRNN, infer is done by calling the `eval` function, which run the whole main_program, see

https://github.com/PaddlePaddle/benchmark/blob/2ce1cea026ed08a3d7d75cfec554000197c54c3e/PaddingRNN/lstm_paddle/train.py#L222-L224

I add profiler and the profiling result is:

```
------------------------->     Profiling Report     <-------------------------

Place: All
Time unit: ms
Sorted by total time in descending order in the same thread

Event                                             Calls       Total       CPU Time (Ratio)        GPU Time (Ratio)        Min.        Max.        Ave.        Ratio.      
thread4::sum                                      84650       21461.1     8349.927323 (0.389073)  13111.131332 (0.610927) 0.017111    51.2725     0.253527    0.135774    
thread4::matmul_grad                              41618       16776.4     6057.403447 (0.361068)  10718.966936 (0.638932) 0.305228    46.1497     0.403104    0.106136    
thread4::elementwise_mul                          127175      10195.1     9605.384625 (0.942154)  589.748053 (0.057846)   0.022776    47.4315     0.0801662   0.0644996   
thread4::sigmoid                                  123379      9767.73     9599.024448 (0.982728)  168.706915 (0.017272)   0.019489    42.4027     0.0791685   0.0617956   
thread4::matmul                                   41635       8133.08     3382.738130 (0.415923)  4750.341290 (0.584077)  0.143605    57.3483     0.195342    0.051454    
thread4::elementwise_mul_grad                     123140      7703.39     7460.839502 (0.968514)  242.547343 (0.031486)   0.020012    37.966      0.062558    0.0487355   
thread4::sigmoid_grad                             122858      7671.23     7494.863924 (0.977010)  176.363323 (0.022990)   0.016008    49.527      0.0624398   0.0485321   
thread4::concat                                   83488       7348.26     7175.165854 (0.976445)  173.090984 (0.023555)   0.030968    37.7151     0.0880157   0.0464888   
thread4::GpuMemcpyAsync:CPU->GPU                  166715      6869.92     6587.384449 (0.958874)  282.534227 (0.041126)   0.009833    47.2699     0.0412076   0.0434626   
thread4::elementwise_add                          83003       5825.08     5637.050692 (0.967721)  188.024904 (0.032279)   0.027057    41.137      0.0701791   0.0368524   
thread4::tanh_grad                                81564       5568.56     5453.796672 (0.979391)  114.762548 (0.020609)   0.017168    45.1199     0.0682723   0.0352295   
thread4::elementwise_add_grad                     83217       4787.64     4223.705667 (0.882211)  563.933345 (0.117789)   0.018223    37.1795     0.057532    0.030289    
thread4::split                                    40820       4494.7      4418.452653 (0.983037)  76.245640 (0.016963)    0.045822    44.8442     0.11011     0.0284357   
thread4::reshape2                                 48122       4090.44     4090.373350 (0.999985)  0.061925 (0.000015)     0.0241      39.928      0.0850014   0.0258782   
thread4::tanh                                     82330       3966.94     3847.118451 (0.969796)  119.818361 (0.030204)   0.015481    37.0434     0.0481834   0.0250969   
thread4::concat_grad                              41831       3809.56     3716.392950 (0.975545)  93.164534 (0.024455)    0.03658     36.6025     0.0910702   0.0241012   
thread4::reshape2_grad                            45494       3662.18     3662.134532 (0.999986)  0.050150 (0.000014)     0.024195    42.6774     0.0804982   0.0231688   
thread4::TensorCopy:GPU->GPU                      45494       3146.06     3146.020736 (0.999987)  0.040814 (0.000013)     0.01754     42.661      0.0691533   0.0199036   
thread4::GpuMemcpyAsync(same_gpu):GPU->GPU        48718       3033.86     2963.257005 (0.976727)  70.606335 (0.023273)    0.014971    39.8955     0.062274    0.0191938   
thread4::GpuMemcpySync(same_gpu):GPU->GPU         45494       3014.12     2937.055181 (0.974432)  77.065672 (0.025568)    0.016063    42.6582     0.0662532   0.0190689   
thread4::dropout                                  41833       2789.86     2582.230099 (0.925578)  207.627468 (0.074422)   0.028126    32.6112     0.0666904   0.0176501   
thread4::dropout_grad                             41502       2665.41     2595.460415 (0.973755)  69.954007 (0.026245)    0.02188     45.5689     0.0642238   0.0168628   
thread4::reduce_sum                               4805        2595.82     2372.601794 (0.914008)  223.221312 (0.085992)   0.052183    47.8319     0.540234    0.0164225   
thread4::fill_zeros_like                          41590       2344.2      2289.563291 (0.976693)  54.636760 (0.023307)    0.015119    39.1325     0.0563645   0.0148306   
thread4::slice_grad                               22922       1449.45     1382.377407 (0.953728)  67.068187 (0.046272)    0.022176    43.1482     0.0632338   0.00916993  
thread4::slice                                    22897       1386.03     1337.307369 (0.964844)  48.727538 (0.035156)    0.018629    32.291      0.0605335   0.00876876  
thread4::square                                   4029        1022.97     622.611856 (0.608631)   400.359040 (0.391369)   0.017343    43.4815     0.253902    0.00647183  
thread4::sgd                                      4039        890.69      328.326876 (0.368621)   562.363516 (0.631379)   0.017136    40.9701     0.220523    0.00563496  
thread4::lookup_table                             606         597.077     592.770617 (0.992788)   4.306112 (0.007212)     0.048315    53.4801     0.985275    0.00377741  
thread4::TensorCopy:GPU->CPU                      1774        165.135     165.064040 (0.999573)   0.070535 (0.000427)     0.033234    21.4063     0.093086    0.00104473  
thread4::GpuMemcpySync:GPU->CPU                   1774        160.129     155.991392 (0.974163)   4.137174 (0.025837)     0.031859    21.4047     0.0902641   0.00101305  
thread4::transpose2                               1723        125.645     121.785104 (0.969281)   3.859618 (0.030719)     0.025357    0.341108    0.0729221   0.000794892 
thread4::lookup_table_grad                        594         123.367     77.969737 (0.632015)    45.397132 (0.367985)    0.120001    11.1448     0.207688    0.000780482 
thread4::softmax_with_cross_entropy_grad          614         77.0217     74.081395 (0.961825)    2.940327 (0.038175)     0.049609    34.1446     0.125443    0.000487279 
thread4::reduce_sum_grad                          599         64.6446     62.817764 (0.971740)    1.826873 (0.028260)     0.060063    0.507589    0.107921    0.000408975 
thread4::softmax_with_cross_entropy               561         63.9663     55.354595 (0.865372)    8.611678 (0.134628)     0.076631    5.56441     0.114022    0.000404683 
thread4::transpose2_grad                          613         38.196      36.399253 (0.952959)    1.796765 (0.047041)     0.030436    0.180926    0.06231     0.000241647 
thread4::reduce_mean                              596         36.2214     36.167686 (0.998518)    0.053694 (0.001482)     0.041292    0.146046    0.0607741   0.000229155 
thread4::reduce_mean_grad                         523         34.2843     32.937570 (0.960720)    1.346689 (0.039280)     0.0394      2.23881     0.0655531   0.0002169   
thread4::elementwise_max                          607         31.4111     30.160321 (0.960182)    1.250734 (0.039818)     0.035725    0.121163    0.051748    0.000198722 
thread4::elementwise_div                          629         30.1106     28.773268 (0.955585)    1.337379 (0.044415)     0.030203    0.126751    0.0478707   0.000190495 
thread4::fill_constant                            616         27.9332     27.172000 (0.972750)    0.761175 (0.027250)     0.026139    0.239362    0.0453461   0.000176719 
thread4::sqrt                                     584         20.1853     19.000572 (0.941306)    1.184763 (0.058694)     0.021661    0.125822    0.0345639   0.000127703 
thread3::sum                                      84646       21018.7     8272.309101 (0.393569)  12746.412801 (0.606431) 0.017155    47.0683     0.248313    0.13168     
thread3::matmul_grad                              41948       16865.5     6068.784678 (0.359834)  10796.735822 (0.640166) 0.306423    43.151      0.402058    0.105661    
thread3::elementwise_mul                          128132      10323       9717.889686 (0.941380)  605.135271 (0.058620)   0.022918    54.8292     0.0805655   0.0646729   
thread3::sigmoid                                  123847      9739.53     9570.225866 (0.982617)  169.299653 (0.017383)   0.019262    41.3597     0.0786416   0.0610173   
thread3::matmul                                   42083       8248.82     3449.409318 (0.418170)  4799.406762 (0.581830)  0.143482    44.9558     0.196013    0.0516781   
thread3::concat                                   84568       7801.67     7626.001600 (0.977483)  175.667600 (0.022517)   0.033232    89.0755     0.0922532   0.0488768   
thread3::elementwise_mul_grad                     123577      7716.15     7472.792451 (0.968461)  243.356583 (0.031539)   0.019165    50.6299     0.06244     0.048341    
thread3::sigmoid_grad                             123203      7591.22     7414.290948 (0.976693)  176.931052 (0.023307)   0.016035    49.644      0.0616156   0.0475583   
thread3::GpuMemcpyAsync:CPU->GPU                  167712      7270.44     6986.389441 (0.960931)  284.051913 (0.039069)   0.009944    88.906      0.0433508   0.0455487   
thread3::elementwise_add                          83413       6101.3      5912.764009 (0.969099)  188.536731 (0.030901)   0.026725    41.5743     0.0731457   0.0382241   
thread3::tanh_grad                                82219       5329.77     5214.098122 (0.978296)  115.675910 (0.021704)   0.016817    42.4125     0.0648241   0.0333906   
thread3::elementwise_add_grad                     82694       4872.37     4315.720219 (0.885754)  556.650148 (0.114246)   0.018258    39.0028     0.0589205   0.030525    
thread3::split                                    40984       4531.12     4454.604018 (0.983113)  76.517391 (0.016887)    0.046588    57.5045     0.110558    0.0283871   
thread3::reshape2                                 48484       4506.94     4506.851323 (0.999980)  0.091994 (0.000020)     0.026099    53.3801     0.0929573   0.0282356   
thread3::tanh                                     82246       3852.08     3732.215688 (0.968883)  119.863718 (0.031117)   0.015346    35.9005     0.0468361   0.0241329   
thread3::concat_grad                              41560       3769.47     3677.195131 (0.975519)  92.279702 (0.024481)    0.037337    44.74       0.0906996   0.0236154   
thread3::reshape2_grad                            46017       3696.72     3696.665546 (0.999986)  0.051671 (0.000014)     0.024031    42.7939     0.0803337   0.0231596   
thread3::GpuMemcpyAsync(same_gpu):GPU->GPU        49099       3447.22     3376.070399 (0.979359)  71.154510 (0.020641)    0.014627    53.2735     0.0702097   0.0215966   
thread3::TensorCopy:GPU->GPU                      46017       3171.38     3171.355163 (0.999991)  0.029248 (0.000009)     0.017832    42.7786     0.0689177   0.0198684   
thread3::GpuMemcpySync(same_gpu):GPU->GPU         46017       3038.52     2960.461020 (0.974311)  78.056497 (0.025689)    0.016696    42.7756     0.0660303   0.019036    
thread3::dropout_grad                             42252       2880.86     2809.613972 (0.975269)  71.246216 (0.024731)    0.022207    42.1285     0.0681828   0.0180483   
thread3::dropout                                  41822       2665.13     2457.060553 (0.921929)  208.069361 (0.078071)   0.029697    32.3827     0.0637255   0.0166968   
thread3::reduce_sum                               4757        2522.73     2300.529435 (0.911919)  222.204673 (0.088081)   0.052907    40.3105     0.53032     0.0158047   
thread3::fill_zeros_like                          41873       2333.12     2278.295352 (0.976503)  54.822114 (0.023497)    0.015066    41.8931     0.0557189   0.0146168   
thread3::slice_grad                               23188       1530.28     1462.223778 (0.955528)  68.054672 (0.044472)    0.023619    39.1938     0.0659944   0.00958706  
thread3::slice                                    22957       1460.65     1411.729807 (0.966507)  48.921364 (0.033493)    0.01877     32.584      0.0636255   0.00915085  
thread3::square                                   4169        1050.81     631.697578 (0.601155)   419.108310 (0.398845)   0.019328    46.8102     0.252052    0.00658321  
thread3::sgd                                      4154        874.429     288.395718 (0.329810)   586.033744 (0.670190)   0.017463    41.0173     0.210503    0.00547822  
thread3::lookup_table                             568         449.338     445.242166 (0.990884)   4.096098 (0.009116)     0.047774    47.1767     0.791088    0.00281507  
thread3::TensorCopy:GPU->CPU                      1722        143.743     143.688523 (0.999620)   0.054618 (0.000380)     0.035987    1.73464     0.0834745   0.000900538 
thread3::GpuMemcpySync:GPU->CPU                   1722        138.922     134.907628 (0.971102)   4.014557 (0.028898)     0.035038    1.73232     0.0806749   0.000870335 
thread3::transpose2                               1788        132.529     128.516475 (0.969721)   4.012835 (0.030279)     0.023066    1.95579     0.0741215   0.000830285 
thread3::lookup_table_grad                        581         107.86      63.386260 (0.587671)    44.473923 (0.412329)    0.12021     1.18059     0.185646    0.000675735 
thread3::softmax_with_cross_entropy               643         94.7732     84.921860 (0.896054)    9.851312 (0.103946)     0.078553    26.8291     0.147392    0.000593746 
thread3::reduce_sum_grad                          608         65.0244     63.215507 (0.972182)    1.808875 (0.027818)     0.060324    0.416985    0.106948    0.000407372 
thread3::fill_constant                            593         45.6734     44.939224 (0.983925)    0.734181 (0.016075)     0.025828    19.6169     0.0770209   0.00028614  
thread3::softmax_with_cross_entropy_grad          597         41.6845     38.831124 (0.931547)    2.853420 (0.068453)     0.049603    0.224853    0.0698234   0.00026115  
thread3::reduce_mean                              615         37.948      37.899799 (0.998728)    0.048251 (0.001272)     0.042479    0.143141    0.0617041   0.000237741 
thread3::reduce_mean_grad                         579         36.0295     34.534294 (0.958499)    1.495248 (0.041501)     0.036602    0.244923    0.0622272   0.000225722 
thread3::transpose2_grad                          585         35.745      34.025293 (0.951889)    1.719745 (0.048111)     0.030572    0.164276    0.0611026   0.00022394  
thread3::elementwise_max                          582         30.8991     29.713066 (0.961615)    1.186055 (0.038385)     0.037489    0.118137    0.0530913   0.00019358  
thread3::elementwise_div                          595         28.4774     27.337505 (0.959971)    1.139930 (0.040029)     0.029853    0.124506    0.0478612   0.000178409 
thread3::sqrt                                     602         20.4824     19.415768 (0.947926)    1.066593 (0.052074)     0.022825    0.073668    0.0340239   0.00012832  
thread2::sum                                      84934       20247.6     8170.576014 (0.403534)  12076.987239 (0.596466) 0.0174      61.6241     0.238392    0.127284    
thread2::matmul_grad                              41534       16637       5952.392701 (0.357780)  10684.607153 (0.642220) 0.307617    43.642      0.400563    0.104586    
thread2::elementwise_mul                          127351      10071.8     9472.425362 (0.940492)  599.352275 (0.059508)   0.022672    42.222      0.0790868   0.063315    
thread2::sigmoid                                  123519      9769.77     9600.938040 (0.982719)  168.832996 (0.017281)   0.019033    42.3149     0.0790953   0.0614165   
thread2::matmul                                   41525       8286.4      3549.433453 (0.428345)  4736.963072 (0.571655)  0.143724    42.4532     0.199552    0.0520914   
thread2::concat                                   84335       8053.25     7878.206673 (0.978265)  175.039482 (0.021735)   0.032306    47.1645     0.0954912   0.0506258   
thread2::elementwise_mul_grad                     123195      7942.6      7699.882508 (0.969441)  242.719008 (0.030559)   0.019793    41.7254     0.0644718   0.0499302   
thread2::sigmoid_grad                             123616      7912.1      7734.617939 (0.977569)  177.479390 (0.022431)   0.016384    45.669      0.0640054   0.0497384   
thread2::GpuMemcpyAsync:CPU->GPU                  167712      7208.13     6923.979874 (0.960579)  284.148856 (0.039421)   0.009741    47.0493     0.0429792   0.045313    
thread2::elementwise_add                          82685       6019.66     5832.305587 (0.968876)  187.355245 (0.031124)   0.026479    47.7753     0.0728023   0.0378419   
thread2::tanh_grad                                82037       5223.54     5108.237802 (0.977927)  115.297488 (0.022073)   0.01722     33.379      0.0636729   0.0328371   
thread2::elementwise_add_grad                     82979       4965.99     4403.756435 (0.886783)  562.233869 (0.113217)   0.01764     45.9351     0.0598464   0.0312181   
thread2::reshape2                                 48135       4499.15     4499.107609 (0.999990)  0.043270 (0.000010)     0.023701    45.4637     0.0934694   0.0282834   
thread2::split                                    41097       4493.93     4417.315871 (0.982951)  76.619086 (0.017049)    0.044816    46.9013     0.109349    0.0282506   
thread2::concat_grad                              41710       3955.03     3862.336965 (0.976564)  92.689670 (0.023436)    0.037319    41.9495     0.094822    0.0248628   
thread2::tanh                                     81895       3908.8      3789.619542 (0.969509)  119.183623 (0.030491)   0.015515    45.2861     0.0477294   0.0245722   
thread2::reshape2_grad                            45524       3640.59     3640.544989 (0.999987)  0.048727 (0.000013)     0.023695    38.2055     0.0799709   0.0228862   
thread2::GpuMemcpyAsync(same_gpu):GPU->GPU        48712       3440.18     3369.795502 (0.979539)  70.389404 (0.020461)    0.01507     45.3872     0.0706229   0.0216263   
thread2::TensorCopy:GPU->GPU                      45524       3122.83     3122.790454 (0.999988)  0.038712 (0.000012)     0.01672     38.1921     0.0685974   0.0196313   
thread2::GpuMemcpySync(same_gpu):GPU->GPU         45524       2994.39     2917.239191 (0.974235)  77.151666 (0.025765)    0.01523     38.1876     0.0657761   0.0188239   
thread2::dropout                                  41538       2689.14     2482.745446 (0.923250)  206.392496 (0.076750)   0.0282      37.0303     0.0647392   0.0169049   
thread2::dropout_grad                             41427       2623.58     2553.716537 (0.973370)  69.864989 (0.026630)    0.021045    38.016      0.0633302   0.0164928   
thread2::reduce_sum                               4644        2462.91     2252.980908 (0.914763)  209.930752 (0.085237)   0.052228    34.8755     0.530343    0.0154828   
thread2::fill_zeros_like                          41864       2425.69     2370.860773 (0.977396)  54.829285 (0.022604)    0.014878    38.9652     0.0579421   0.0152488   
thread2::slice                                    23001       1546.19     1497.254033 (0.968351)  48.935482 (0.031649)    0.018163    47.1092     0.0672227   0.00971993  
thread2::slice_grad                               22986       1376.39     1308.860779 (0.950935)  67.533006 (0.049065)    0.020793    26.3624     0.0598797   0.00865253  
…